Add fetchVMAgentTargets to wrap the functional behavior of getting some target info. Add currently-failing unit test for what it _should_ return, along with a helper to pre-populate a req and ship back some golden data

+++ main.go
@@ -13,7 +13,7 @@ type VMAgentAPITarget struct {
 	ScrapeURL          string            `json:"scrapeUrl"`
 	LastError          string            `json:"lastError"`
 	LastScrape         time.Time         `json:"lastScrape"`
-	LastScrapeDuration int               `json:"lastScrapeDuration"`
+	LastScrapeDuration float64           `json:"lastScrapeDuration"`
 	LastSamplesScraped int               `json:"lastSamplesScraped"`
 	Health             string            `json:"health"`
 }
